合并了修改版本号4450, 4556-4557, 4789, 4827, 4872, 4892, 4908, 4913, 4937, 4939, 5008-5009, 5165, 5169-5171, 5178, 5186, 5195, 5235, 5237, 5239, 5245, 5247, 5263, 5290, 5306, 5320, 5329 从 branches/CodeTest: 连锁掌柜-后台-优惠券 新建优惠券时可以绑定分类/分组/sku/商品 连锁掌柜-后台-游戏活动 添加奖品判断奖品的数量是否超过未使用优惠券数量 连锁掌柜-前端-游戏活动  奖励优惠券时注明优惠券来源 连锁掌柜-前端-优惠券转赠 修改判断条件 优惠券可以转赠N次 ........ 合并了修改版本号4554 从 branches/Code: 连锁掌柜-后台-门店 导入的错误信息修改 连锁掌柜-后台-门店 导入的错误信息修改
........

........
合并了修改版本号4555 从 branches/Code:
连锁掌柜-后台-优惠券  判断是否有判断sku/分类/分组 
........

........
合并了修改版本号4775 从 branches/Code:
连锁掌柜-后台-优惠券 增加优惠券绑定
连锁掌柜-后台-异业合作 奖励模板 增加销售奖励
........

........
合并了修改版本号4826 从 branches/Code:
连锁掌柜-后台-异业合作  奖励模版增加商品奖励栏目
........

........
合并了修改版本号4837, 4844, 4855-4856, 4862, 4868 从 branches/Code:
连锁掌柜-后台-异业合作  奖励模版修改


........
合并了修改版本号4875, 4885, 4891 从 branches/Code:
连锁掌柜-后台-游戏活动  增加问卷调查功能
........
连锁掌柜-后台-游戏活动  增加问卷调查功能  默认值 状态status=0
........
连锁掌柜-后台-触点活动 追加奖品数量时同时生成奖品池数量
........

........
合并了修改版本号4900, 4907 从 branches/Code:
连锁掌柜-后台-异业合作  更新分销商的合作类型
........
连锁掌柜-前端-商城  商品列表判断是否小店
........

........
合并了修改版本号4912 从 branches/Code:
连锁掌柜-前端-商城  商品列表判断是否小店
连锁掌柜-后台-游戏活动 增加花样问卷
........

........
连锁掌柜-前台-分销商 获取商品sql变更

........
合并了修改版本号4938 从 branches/Code:
连锁掌柜-前台-分销商 获取商品sql变更,商品类增加IsCheck

........

........
连锁掌柜-前台-分销设置  获取分销商品加上IsDelete状态
........
连锁掌柜-前台-分销设置  获取分销商品加上IsDelete状态
........
合并了修改版本号5164 从 branches/Code:
连锁掌柜-前端-游戏活动  奖励接口扣除积分逻辑修改为有参加就扣
........

........
合并了修改版本号5166 从 branches/Code:
连锁掌柜-后台-异业合作  新增销售奖励  
........

........
合并了修改版本号5168 从 branches/Code:
连锁掌柜-前台-游戏活动  获取游戏活动信息背景图片 增加游戏规则
........

........
合并了修改版本号5167 从 branches/Code:
连锁掌柜-后台-产品  产品列表去除当面付
........

........
合并了修改版本号5177 从 branches/Code:
连锁掌柜-后台-异业合作 分销商编辑
........

........
连锁掌柜-前台-异业合作 分销商15天会员变化情况
........
连锁掌柜-前台-保存分销商    增加SalesType字段
........
连锁掌柜-前台- 分销商  销售额接口  分销商品
........
连锁掌柜-前台- 分销商    分销商品
........
合并了修改版本号5238 从 branches/Code:
连锁掌柜-前台- 分销商    分销商品
........

........
合并了修改版本号5244 从 branches/Code:
连锁掌柜-前台-分销商    获取商品
........

........
合并了修改版本号5246 从 branches/Code:
连锁掌柜-前台-分销商    获取商品SQL语句
........

........
合并了修改版本号5261 从 branches/Code:
连锁掌柜-后台-分销商管理    修改分销商体现获取的表 VipWithdrawDeposit
........

........
合并了修改版本号5287 从 branches/Code:
连锁掌柜-前端-分销商 分销商销售情况
........

........
合并了修改版本号5305 从 branches/Code:
连锁掌柜-前端-分销商 每日奖励情况 增加销售分类
........

........
连锁掌柜-后台-商城 获取活动列表时返回总页数
........
合并了修改版本号5328 从 branches/Code:
连锁掌柜-促销活动-活动设置  增加RuleContent
........

........


git-svn-id: https://10.133.192.229/svn/CustomerManage/trunk/Code@5348 67192a19-bcbd-5648-baee-99b7ca3a6e28

+++ CPOS.BS.DataAccess/CouponDAO.cs
@@ -714,7 +714,7 @@ namespace JIT.CPOS.BS.DataAccess
         public int GetCouponCountByCouponTypeID(string strCouponTypeId)
         {
             string strSql = string.Format("select  count(1)CouponCount from Coupon a WITH(NOLOCK) LEFT join VipCouponMapping b WITH(NOLOCK) ON a.CouponID=b.CouponID WHERE   a.IsDelete = 0 AND a.[Status] = 0 and  b.VIPID is null and a.CouponTypeID='{0}'", strCouponTypeId);
-            return Convert.ToInt16(this.SQLHelper.ExecuteScalar(strSql));
+            return Convert.ToInt32(this.SQLHelper.ExecuteScalar(strSql));
         }
      }
 }
